requirements.txt
Pillow
PyMuPDF
beautifulsoup4
markdown
ollama
python-docx
